Split view methods out of the Test class.
